♻️ move getPathToNestedCSSRule to styleSheetObserver
