+++ packages/rum/src/domain/record/observers/styleSheetObserver.ts
@@ -1,7 +1,6 @@
 import { instrumentMethodAndCallOriginal } from '@datadog/browser-core'
 import type { StyleSheetRule } from '../../../types'
 import { getSerializedNodeId, hasSerializedNode } from '../serialization'
-import { getPathToNestedCSSRule } from '../utils'
 import type { ListenerHandler } from './utils'
 
 type GroupingCSSRuleTypes = typeof CSSGroupingRule | typeof CSSMediaRule | typeof CSSSupportsRule
@@ -63,4 +62,25 @@ export function initStyleSheetObserver(cb: StyleSheetCallback): ListenerHandler 
   }
 
   return () => instrumentationStoppers.forEach((stopper) => stopper.stop())
+}
+
+export function getPathToNestedCSSRule(rule: CSSRule): number[] | undefined {
+  const path: number[] = []
+  let currentRule = rule
+  while (currentRule.parentRule) {
+    const rules = Array.from((currentRule.parentRule as CSSGroupingRule).cssRules)
+    const index = rules.indexOf(currentRule)
+    path.unshift(index)
+    currentRule = currentRule.parentRule
+  }
+  // A rule may not be attached to a stylesheet
+  if (!currentRule.parentStyleSheet) {
+    return
+  }
+
+  const rules = Array.from(currentRule.parentStyleSheet.cssRules)
+  const index = rules.indexOf(currentRule)
+  path.unshift(index)
+
+  return path
 }
