+++ packages/rum/src/domain/record/utils.ts
@@ -21,24 +21,3 @@ export function assembleIncrementalSnapshot<Data extends BrowserIncrementalData>
     timestamp: timeStampNow(),
   }
 }
-
-export function getPathToNestedCSSRule(rule: CSSRule): number[] | undefined {
-  const path: number[] = []
-  let currentRule = rule
-  while (currentRule.parentRule) {
-    const rules = Array.from((currentRule.parentRule as CSSGroupingRule).cssRules)
-    const index = rules.indexOf(currentRule)
-    path.unshift(index)
-    currentRule = currentRule.parentRule
-  }
-  // A rule may not be attached to a stylesheet
-  if (!currentRule.parentStyleSheet) {
-    return
-  }
-
-  const rules = Array.from(currentRule.parentStyleSheet.cssRules)
-  const index = rules.indexOf(currentRule)
-  path.unshift(index)
-
-  return path
-}
